+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Avoid mosaic layout when screensharing is enabled.
+- Video conference stabilization.
+
 ### Changed
 - Update SDK to 5.4.50